requirements.txt
Django==5.2.8
psycopg[binary]
dj-database-url
weasyprint==63.0
django-admin-rangefilter
django-unfold
django-simple-history
amadeus>=9.0.0
requests
pytest-django